requirement.txt
django==2.0.7
pymysql==0.9.3
djangorestframework==3.9.2
django_filter==2.1.0
django_redis==4.10.0
redis==2.10.6
django-cors-headers==2.4.0
django-redis-cache==1.8.1
